Restructure parsers into sections such as tokens, combinators, transformers, helpers ,etc
